refactored all but upsert to have a safe adapter
