+++ internal/cloudformation/tasks/auth.go
