+++ internal/cloudformation/tasks/util.go
@@ -1,14 +1,11 @@
 package tasks
 
 import (
-	"fmt"
 	"os"
 	"strings"
 
 	log "github.com/sirupsen/logrus"
 
-	"github.com/aws/aws-sdk-go/aws"
-	awsCF "github.com/aws/aws-sdk-go/service/cloudformation"
 	"github.com/urfave/cli"
 )
 
@@ -49,30 +46,4 @@ func getParamMap(c *cli.Context) map[string]string {
 		}
 	}
 	return paramMap
-}
-
-func printStackEvents(cf *awsCF.CloudFormation, stackName string) {
-	status, err := cf.DescribeStackEvents(&awsCF.DescribeStackEventsInput{StackName: aws.String(stackName)})
-	checkError(err)
-
-	fmt.Println()
-	fmt.Printf(" %-19v | %-22v | %-30v | %v | %v | \n", "Time", "Status", "Type", "LogicalID", "Status Reason")
-	for _, event := range status.StackEvents {
-		if event.Timestamp != nil {
-			fmt.Printf(" %-19v |", event.Timestamp.Format("2006-01-2 15:04:05"))
-		}
-		if event.ResourceStatus != nil {
-			fmt.Printf(" %-22v |", *event.ResourceStatus)
-		}
-		if event.ResourceType != nil {
-			fmt.Printf(" %-30v |", *event.ResourceType)
-		}
-		if event.LogicalResourceId != nil {
-			fmt.Printf(" %v |", *event.LogicalResourceId)
-		}
-		if event.ResourceStatusReason != nil {
-			fmt.Printf(" %v |", *event.ResourceStatusReason)
-		}
-		fmt.Println()
-	}
 }
