+++ internal/tasks/events.go
@@ -1,19 +1,13 @@
 package tasks
 
 import (
+	"github.com/KablamoOSS/kombustion/internal/cloudformation/tasks"
 	"github.com/urfave/cli"
 )
 
-var PrintEventsFlags = []cli.Flag{
-	cli.StringFlag{
-		Name:  "region, r",
-		Usage: "region to deploy to",
-		Value: "ap-southeast-2",
-	},
-}
-
+// PrintStackEvents outputs the events of a stack
 func PrintEvents(c *cli.Context) {
 	cf := getCF(c.GlobalString("profile"), c.String("region"))
 	stackName := c.Args().Get(0)
-	printStackEvents(cf, stackName)
+	tasks.PrintStackEvents(cf, stackName)
 }
