+++ internal/tasks/generate.go
@@ -1,0 +1,35 @@
+package tasks
+
+import (
+	"github.com/KablamoOSS/kombustion/internal/cloudformation/tasks"
+	"github.com/urfave/cli"
+)
+
+// GenerateFlags for kombustion upsert
+var GenerateFlags = []cli.Flag{
+	cli.StringSliceFlag{
+		Name:  "param, p",
+		Usage: "cloudformation parameters. eg. ( --param Env=dev --param BucketName=test )",
+	},
+	cli.BoolFlag{
+		Name:  "no-base-outputs, b",
+		Usage: "disable generation of outputs for Base AWS types",
+	},
+}
+
+func init() {
+	GenerateFlags = append(CloudFormationStackFlags)
+}
+
+// Generate a template and save it to disk, without upserting it
+func Generate(c *cli.Context) {
+	paramMap := getParamMap(c)
+
+	tasks.GenerateTemplate(cloudformation.GenerateParams{
+		Filename:           c.Args().Get(0),
+		EnvFile:            c.String("env-file"),
+		Env:                c.String("env"),
+		DisableBaseOutputs: c.Bool("no-base-outputs"),
+		ParamMap:           paramMap,
+	}))
+}
